Extract applyPatch function used in InMemoryCache and RedisCache
